+++ layouts/default.vue
@@ -1,5 +1,5 @@
 <template>
-    <div class="app" :class="screenClass" v-loading="loading">
+    <div class="app" :class="screenClass">
         <Menu v-if="!isLoginPage"/>
         <el-row class="main">
             <el-col :span="showIntro ? 16 : 24" class="view-wrapper">
@@ -24,14 +24,11 @@
     components: {Menu, Intro, Footer},
   })
   export default class App extends Vue {
-    get loading() {
-      return this.$store.state.loading
-    }
     get screenClass(): string {
-      return this.$route.name === "Login" ? "full-screen" : "normal-screen"
+      return this.$route.name === "index.vue" ? "full-screen" : "normal-screen"
     }
     get isLoginPage(): boolean {
-      return this.$route.name === "Login"
+      return this.$route.name === "index.vue"
     }
     get showIntro() {
       return !this.isLoginPage && this.$store.state.auth.isLogin
